+++ example-library/src/main/java/dk/ilios/example/library/LibraryModel.java
@@ -1,0 +1,7 @@
+package dk.ilios.example.library;
+
+import io.realm.RealmObject;
+
+public class LibraryModel extends RealmObject {
+    public String libraryField;
+}