+++ example-library/src/main/java/dk/ilios/example/library/LibraryModule.java
@@ -1,0 +1,7 @@
+package dk.ilios.example.library;
+
+import io.realm.annotations.RealmModule;
+
+@RealmModule(library = true, allClasses = true)
+public class LibraryModule {
+}