+++ example/src/main/java/dk/ilios/example/realmfieldnames/AppModule.java
@@ -1,0 +1,7 @@
+package dk.ilios.example.realmfieldnames;
+
+import io.realm.annotations.RealmModule;
+
+@RealmModule(library = false, allClasses = true)
+public class AppModule {
+}