+++ library/src/main/kotlin/dk/ilios/realmfieldnames/ClassData.kt
@@ -5,16 +5,12 @@ import java.util.TreeMap
 /**
  * Class responsible for keeping track of the metadata for each Realm model class.
  */
-class ClassData(val packageName: String?, val simpleClassName: String) {
+class ClassData(val packageName: String?, val simpleClassName: String, val libraryClass: Boolean = false) {
 
     val fields = TreeMap<String, String?>() // <fieldName, linkedType or null>
 
     fun addField(field: String, linkedType: String?) {
         fields.put(field, linkedType)
-    }
-
-    fun getFields(): Map<String, String?> {
-        return fields
     }
 
     val qualifiedClassName: String