+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.1.1
 - Fixed a bug that crashed the annotation processor if Realm wasn't on the classpath (#14).
 - Fixed crash when processing Kotlin model classes (#9).
+- Fixed field names not being generated for library class references (#15).
 
 ### 1.1.0
  - Fixed a bug where hungarian notation was detected in the middle of a String.